CU-51r1cy Add test and google auth tool for backup.

+++ hive/main/hive_backup.py
@@ -3,8 +3,9 @@ import pathlib
 import subprocess
 
 from hive.settings import VAULTS_BASE_DIR
+from hive.util.auth import did_auth
 from hive.util.common import did_tail_part, create_full_path_dir
-from hive.util.constants import APP_ID, VAULT_ACCESS_R
+from hive.util.constants import APP_ID, VAULT_ACCESS_R, HIVE_MODE_TEST, HIVE_MODE_DEV
 from hive.util.did_info import get_all_did_info_by_did
 from hive.util.did_mongo_db_resource import export_mongo_db, import_mongo_db, delete_mongo_db_backup
 from hive.util.vault_backup_info import *
@@ -14,12 +15,15 @@ from hive.main.interceptor import post_json_param_pre_proc
 
 
 class HiveBackup:
+    mode = HIVE_MODE_DEV
+
     def __init__(self):
         self.app = None
         self.response = ServerResponse("HiveBackup")
 
-    def init_app(self, app):
+    def init_app(self, app, mode):
         self.app = app
+        HiveBackup.mode = mode
 
     def __proc_google_drive_param(self):
         did, app_id, content, err = post_json_param_pre_proc(self.response,
@@ -46,23 +50,29 @@ class HiveBackup:
         return did, drive_name, None
 
     def save_to_google_drive(self):
-        did, drive_name, err = self.__proc_google_drive_param()
-        if err:
-            return err
-
-        _thread.start_new_thread(HiveBackup.save_vault_data, (did, drive_name))
-
+        did, drive_name, response = self.__proc_google_drive_param()
+        if response:
+            return response
+        if HiveBackup.mode != HIVE_MODE_TEST:
+            _thread.start_new_thread(HiveBackup.save_vault_data, (did, drive_name))
         return self.response.response_ok()
 
     def restore_from_google_drive(self):
         did, drive_name, err = self.__proc_google_drive_param()
         if err:
             return err
-
-        _thread.start_new_thread(HiveBackup.restore_vault_data, (did, drive_name))
+        if HiveBackup.mode != HIVE_MODE_TEST:
+            _thread.start_new_thread(HiveBackup.restore_vault_data, (did, drive_name))
+        return self.response.response_ok()
 
     def get_sync_google_drive_state(self):
-        pass
+        did, app_id = did_auth()
+        if (did is None) or (app_id is None):
+            return self.response.response_err(401, "auth failed")
+
+        info = get_vault_backup_info(did)
+        data = {"hive_backup_info": info}
+        return self.response.response_ok(data)
 
     @staticmethod
     def gene_did_google_drive_name(did):
@@ -98,11 +108,14 @@ class HiveBackup:
 
     @staticmethod
     def restore_vault_data(did, drive_name):
+        rclone_config = RcloneTool.find_rclone_config_file(drive_name)
+        if not rclone_config.exists():
+            return
         update_vault_backup_state(did, VAULT_BACKUP_STATE_RESTORE, VAULT_BACKUP_MSG_SUCCESS)
         did_folder = HiveBackup.get_did_vault_path(did)
         if not did_folder.exists():
             create_full_path_dir(did_folder)
-        line = 'rclone sync %s:elastos_hive_node_data %s' % (drive_name, did_folder.as_posix())
+        line = f'rclone  --config {rclone_config.as_posix()} sync {drive_name}:elastos_hive_node_data {did_folder.as_posix()}'
         subprocess.call(line, shell=True)
         HiveBackup.import_did_mongodb_data(did)
         update_vault_backup_state(did, VAULT_BACKUP_STATE_STOP, VAULT_BACKUP_MSG_SUCCESS)
@@ -111,11 +124,14 @@ class HiveBackup:
 
     @staticmethod
     def save_vault_data(did, drive_name):
+        rclone_config = RcloneTool.find_rclone_config_file(drive_name)
+        if not rclone_config.exists():
+            return
         update_vault_backup_state(did, VAULT_BACKUP_STATE_BACKUP, VAULT_BACKUP_MSG_SUCCESS)
         HiveBackup.export_did_mongodb_data(did)
         did_folder = HiveBackup.get_did_vault_path(did)
-        line = 'rclone sync %s %s:elastos_hive_node_data' % (did_folder.as_posix(), drive_name)
+        line = f'rclone --config {rclone_config.as_posix()} sync {did_folder.as_posix()} {drive_name}:elastos_hive_node_data'
         subprocess.call(line, shell=True)
         update_vault_backup_state(did, VAULT_BACKUP_STATE_STOP, VAULT_BACKUP_MSG_SUCCESS)
-        HiveBackup.delete_did_mongodb_export_data(did)
         RcloneTool.remove_rclone_config_file(drive_name)
+        HiveBackup.delete_did_mongodb_export_data(did)
